requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.4
